refactor: elasticSearch 제거
