+++ build.gradle
@@ -95,8 +95,8 @@ dependencies {
 
     implementation "org.flywaydb:flyway-core"
     implementation "org.postgresql:postgresql:42.7.7"
-	implementation("org.flywaydb:flyway-core")
-	implementation("org.flywaydb:flyway-database-postgresql")
+//	implementation("org.flywaydb:flyway-core")
+//	implementation("org.flywaydb:flyway-database-postgresql")
 }
 
 dependencyManagement {
