+++ src/main/java/core/domain/post/service/RecentSearchRedisService.java
@@ -1,4 +1,4 @@
-package core.global.search.service;
+package core.domain.post.service;
 
 import core.domain.user.entity.User;
 import core.domain.user.repository.UserRepository;
