--- src/main/java/core/global/search/dto/PostCreatedEvent.java
@@ -1,3 +1,0 @@
-package core.global.search.dto;
-
-public record PostCreatedEvent(Long postId, PostDocument doc) {}
