--- src/main/java/core/global/search/dto/PostDeletedEvent.java
@@ -1,3 +1,0 @@
-package core.global.search.dto;
-
-public record PostDeletedEvent(Long postId) {}
