--- src/main/java/core/global/search/dto/PostUpdatedEvent.java
@@ -1,3 +1,0 @@
-package core.global.search.dto;
-
-public record PostUpdatedEvent(Long postId, PostDocument doc) {}
