--- src/main/java/core/global/search/dto/SearchHitLite.java
@@ -1,3 +1,0 @@
-package core.global.search.dto;
-
-public record SearchHitLite(Long id, double score, String highlight) {}
